# docs/openapi/control-plane-fleet-usage.yaml
openapi: 3.0.3
info:
  title: SkuldBot Control Plane - Fleet and Usage API
  version: 1.0.0
  description: |
    Contract for Control Plane communication with tenant Orchestrators.

    Scope:
    - Fleet lifecycle: register, heartbeat, health, deregister.
    - Usage ingestion: batch ingest with idempotency by event_id.
servers:
  - url: https://control-plane.skuld.ai
    description: Production
  - url: https://staging-control-plane.skuld.ai
    description: Staging
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/usage/ingest:
    post:
      tags: [usage]
      summary: Ingest usage batch from orchestrator
      description: |
        Idempotency:
        - Batch-level: unique by (`orchestrator_id`, `batch_id`).
        - Event-level: unique by (`orchestrator_id`, `event_id`).
      parameters:
        - $ref: '#/components/parameters/LicenseKeyHeader'
        - $ref: '#/components/parameters/ApiKeyHeader'
        - $ref: '#/components/parameters/OrchestratorIdHeader'
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/BatchIdHeader'
        - $ref: '#/components/parameters/TraceIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsageBatchRequest'
      responses:
        '200':
          description: Batch processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageBatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Duplicate batch already processed
components:
  securitySchemes:
    FleetTokenHeader:
      type: apiKey
      in: header
      name: x-orchestrator-token
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LicenseKeyHeader:
      name: x-license-key
      in: header
      required: false
      schema:
        type: string
    ApiKeyHeader:
      name: x-api-key
      in: header
      required: false
      schema:
        type: string
    OrchestratorIdHeader:
      name: x-orchestrator-id
      in: header
      required: true
      schema:
        type: string
    TenantIdHeader:
      name: x-tenant-id
      in: header
      required: false
      schema:
        type: string
    BatchIdHeader:
      name: x-batch-id
      in: header
      required: true
      schema:
        type: string
    TraceIdHeader:
      name: x-trace-id
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    UsageEvent:
      type: object
      required: [id, metric, quantity, occurredAt]
      properties:
        id:
          type: string
        metric:
          type: string
        quantity:
          type: number
        occurredAt:
          type: string
          format: date-time
        botId:
          type: string
        runId:
          type: string
        installationId:
          type: string
        metadata:
          type: object
          additionalProperties: true
    UsageBatchRequest:
      type: object
      required: [batchId, tenantId, events, sentAt]
      properties:
        batchId:
          type: string
        tenantId:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/UsageEvent'
        sentAt:
          type: string
          format: date-time
    UsageBatchResponse:
      type: object
      required: [processedCount, duplicateEventCount, batchId, traceId]
      properties:
        processedCount:
          type: integer
        duplicateEventCount:
          type: integer
        batchId:
          type: string
        traceId:
          type: string
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...